+++ lib/reflow.js
@@ -1,8 +1,7 @@
 import threadPool from './thread-pool';
 
 import dynamicRunner from './runner/dynamic-runner';
-import {createContext} from './runner/Module';
-
+import createContext from './runner/create-context';
 
 import executeMatrix from './execute'
 import { evaluateFlow, evaluateSubflow } from './evaluate';
