+++ lib/runner/Module.js
@@ -3,6 +3,7 @@ import has from 'has';
 import path from 'path';
 import { ok } from 'assert';
 import { runInNewContext } from 'vm';
+import createContext from './create-context';
 
 const NativeModules = process.binding('natives');
 
@@ -16,24 +17,6 @@ function isNativeModule(id) {
   return has(NativeModules, id);
 }
 
-// Creates a sandbox so we don't share globals across different runs.
-
-function createContext(passContext = {}) {
-  const context = {
-    Buffer,
-    clearImmediate,
-    clearInterval,
-    clearTimeout,
-    setImmediate,
-    setInterval,
-    setTimeout,
-    console,
-    process,
-    ...passContext,
-  };
-  context.global = context;
-  return context;
-}
 
 // This class should satisfy the Module interface that NodeJS defines in their native module.js
 // implementation.
